+++ immunity_notifications/context_processors.py
@@ -5,8 +5,12 @@ from immunity_notifications import settings as app_settings
 
 def notification_api_settings(request):
     return {
-        'OPENWISP_NOTIFICATIONS_HOST': app_settings.OPENWISP_NOTIFICATIONS_HOST,
-        'OPENWISP_NOTIFICATIONS_SOUND': static(
-            app_settings.OPENWISP_NOTIFICATIONS_SOUND
+        'IMMUNITY
+_NOTIFICATIONS_HOST': app_settings.IMMUNITY
+_NOTIFICATIONS_HOST,
+        'IMMUNITY
+_NOTIFICATIONS_SOUND': static(
+            app_settings.IMMUNITY
+_NOTIFICATIONS_SOUND
         ),
     }