+++ immunity_notifications/settings.py
@@ -5,21 +5,28 @@ from notifications.settings import CONFIG_DEFAULTS
 
 CONFIG_DEFAULTS.update({'USE_JSONFIELD': True})
 
-OPENWISP_NOTIFICATIONS_HOST = getattr(settings, 'OPENWISP_NOTIFICATIONS_HOST', None)
+IMMUNITY
+_NOTIFICATIONS_HOST = getattr(settings, 'IMMUNITY
+_NOTIFICATIONS_HOST', None)
 
-OPENWISP_NOTIFICATIONS_CACHE_TIMEOUT = getattr(
-    settings, 'OPENWISP_NOTIFICATIONS_CACHE_TIMEOUT', 2 * 24 * 60 * 60
+IMMUNITY
+_NOTIFICATIONS_CACHE_TIMEOUT = getattr(
+    settings, 'IMMUNITY
+_NOTIFICATIONS_CACHE_TIMEOUT', 2 * 24 * 60 * 60
 )
 
 IGNORE_ENABLED_ADMIN = getattr(
-    settings, 'OPENWISP_NOTIFICATIONS_IGNORE_ENABLED_ADMIN', []
+    settings, 'IMMUNITY
+_NOTIFICATIONS_IGNORE_ENABLED_ADMIN', []
 )
 POPULATE_PREFERENCES_ON_MIGRATE = getattr(
-    settings, 'OPENWISP_NOTIFICATIONS_POPULATE_PREFERENCES_ON_MIGRATE', True
+    settings, 'IMMUNITY
+_NOTIFICATIONS_POPULATE_PREFERENCES_ON_MIGRATE', True
 )
 NOTIFICATION_STORM_PREVENTION = getattr(
     settings,
-    'OPENWISP_NOTIFICATIONS_NOTIFICATION_STORM_PREVENTION',
+    'IMMUNITY
+_NOTIFICATIONS_NOTIFICATION_STORM_PREVENTION',
     {
         'short_term_time_period': 10,
         'short_term_notification_count': 6,
@@ -31,20 +38,25 @@ NOTIFICATION_STORM_PREVENTION = getattr(
     },
 )
 
-OPENWISP_NOTIFICATIONS_SOUND = getattr(
+IMMUNITY
+_NOTIFICATIONS_SOUND = getattr(
     settings,
-    'OPENWISP_NOTIFICATIONS_SOUND',
+    'IMMUNITY
+_NOTIFICATIONS_SOUND',
     'immunity-notifications/audio/notification_bell.mp3',
 )
 
 # Remove the leading "/static/" here as it will
 # conflict with the "static()" call in context_processors.py.
 # This is done for backward compatibility.
-OPENWISP_NOTIFICATIONS_SOUND = re.sub('^/static/', '', OPENWISP_NOTIFICATIONS_SOUND)
+IMMUNITY
+_NOTIFICATIONS_SOUND = re.sub('^/static/', '', IMMUNITY
+_NOTIFICATIONS_SOUND)
 
 
 def get_config():
-    user_config = getattr(settings, 'OPENWISP_NOTIFICATIONS_CONFIG', {})
+    user_config = getattr(settings, 'IMMUNITY
+_NOTIFICATIONS_CONFIG', {})
     config = CONFIG_DEFAULTS.copy()
     config.update(user_config)
     return config