+++ immunity_notifications/websockets/handlers.py
@@ -16,7 +16,8 @@ def user_in_notification_storm(user):
     """
     A user is affected by notifications storm if any of short term
     or long term check passes. The checks are configured by
-    "OPENWISP_NOTIFICATIONS_NOTIFICATION_STORM_PREVENTION" setting.
+    "IMMUNITY
+_NOTIFICATIONS_NOTIFICATION_STORM_PREVENTION" setting.
     If the user is found to be affected by a notification storm,
     the value of this function is cached for 60 seconds.
     """